Rename subpackage & move test
