+++ tests/test_cons.py
@@ -153,18 +153,6 @@ def test_printCons():
     m.printCons(c)
 
 
-def test_addPiecewiseLinearCons():
-    m = Model()
-
-    xpoints = [1,3,5]
-    ypoints = [1,2,4]
-    x = m.addVar(lb=xpoints[0], ub=xpoints[-1], obj=2)
-    y = m.addVar(lb=-m.infinity(), obj=-3)
-    m.addPiecewiseLinearCons(x,y,xpoints,ypoints)
-    
-    m.optimize()
-    assert m.isEQ(m.getObjVal(), -2) 
-
 @pytest.mark.skip(reason="TODO: test getValsLinear()")
 def test_getValsLinear():
     assert True
